+++ packages/config/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @vercel/router-sdk
 
+## 0.2.0
+
+### Minor Changes
+
+- Add root to experimentalServices to set a service's working directory. ([#15929](https://github.com/vercel/vercel/pull/15929))
+
 ## 0.1.1
 
 ### Patch Changes